# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Role-play API",
        "description": "Документация API для role-play",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/npcs/": {
            "get": {
                "tags": [
                    "NPCs"
                ],
                "summary": "Список NPC",
                "operationId": "dcca645abededa8beb66beb1fe16aaca",
                "parameters": [
                    {
                        "name": "without_paginate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NpcResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/npcs/{npc}/show": {
            "get": {
                "tags": [
                    "NPCs"
                ],
                "summary": "Получить одного NPC",
                "operationId": "14030059ba6093956d32a33c29e9918a",
                "parameters": [
                    {
                        "name": "npc",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NpcShowResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NPC не найден"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/npcs/create": {
            "post": {
                "tags": [
                    "NPCs"
                ],
                "summary": "Создать нового NPC",
                "operationId": "93ab075f1729aadefeaad8087df75c8d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateNpcRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "NPC успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NpcResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/npcs/{npc}/update": {
            "put": {
                "tags": [
                    "NPCs"
                ],
                "summary": "Обновить NPC",
                "operationId": "4f2a8cc27aed125019ef6f88db0fcfc3",
                "parameters": [
                    {
                        "name": "npc",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNpcRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "NPC успешно обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NpcShowResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NPC не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/npcs/{npc}/delete": {
            "delete": {
                "tags": [
                    "NPCs"
                ],
                "summary": "Удалить NPC",
                "operationId": "f57338da8db0bb188a5b283c821ba78b",
                "parameters": [
                    {
                        "name": "npc",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "NPC удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Npc deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NPC не найден"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Список тегов",
                "operationId": "7f99d744240359db0705709ba89b4c5f",
                "parameters": [
                    {
                        "name": "without_paginate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список тегов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/create": {
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Создать тег",
                "operationId": "df7233a0f5ddb19cfa7e9b857517dc45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Тег успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/{tag}/update": {
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Обновить тег",
                "operationId": "fbfd66a746e73b7542503917ded0fdce",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Тег обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Тег не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/{tag}/delete": {
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Удалить тег",
                "operationId": "252ebde3e08b6ab10924da5a791977fa",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тег успешно удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tag deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Тег не найден"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateNpcRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John the Blacksmith"
                    },
                    "title": {
                        "type": "string",
                        "example": "Master Artisan"
                    },
                    "description": {
                        "type": "string",
                        "example": "Cool description"
                    },
                    "image": {
                        "description": "Изображение (jpeg, png, jpg, gif, svg)",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreTagRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "quest"
                    }
                },
                "type": "object"
            },
...
            "UpdateNpcRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John the Blacksmith"
                    },
                    "title": {
                        "type": "string",
                        "example": "Master Artisan"
                    },
                    "description": {
                        "type": "string",
                        "example": "Cool description"
                    },
                    "image": {
                        "description": "Изображение (jpeg, png, jpg, gif, svg)",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateTagRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "npc"
                    }
                },
                "type": "object"
            },
...
            "NpcResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John the Blacksmith"
                    },
                    "title": {
                        "type": "string",
                        "example": "Master Artisan"
                    }
                },
                "type": "object"
            },
            "NpcShowResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    }
                },
                "type": "object"
            },
            "TagResource": {
                "title": "Tag",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "quest"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Bearer authentication using JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}